pyproject.toml
[project]
name = "radio-watermarks"
version = "0.1.0"
description = "Collect and analyze Swedish DAB+ radio now-playing metadata"
requires-python = ">=3.12"
dependencies = [
    "azure-functions>=1.20",
    "azure-data-tables>=12.5",
    "httpx>=0.27",
    "defusedxml>=0.7",
    "websockets>=13",
]

[project.optional-dependencies]
analyze = [
    "pandas>=2.2",
    "rich>=13.7",
]

[project.scripts]
rwm-analyze = "radio_watermarks.analyze:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["radio_watermarks"]
